Added a centralized class for acting bad and added the light sensor and the gyroscope sensors
I added a Bad_Actor class so that all the sensor values are redirected to one central location before being saved.

This allows us to use different poisoning methods like using the preset value or a random value or anything we want simply by adding one method.

To allow the user to choose different modes of data-poisoning, I changed the UI.

I implemented the new way of poisoning data in the light, gyroscope, and barometer sensors. I'm waiting for your opinions before I implement it for the other sensors.

resolves #21 light sensor
resolves #13 gyroscope
resolves #19 barometer
resolves #11 light sensor again